feat: Add ability to conditionally generate the URL for checks
The URL check takes a very long time because it is only available for
the details of each check, which means there has to be an http request
for every single check in order to generate it.  Most of the time (I
think) list is going to be most helpful to generate the IDs to operate
on, and showing the URL every time takes too long.